+++ backend/src/main/java/com/back/domain/transactions/dto/UpdateTransactionRequestDto.java
@@ -3,11 +3,11 @@ package com.back.domain.transactions.dto;
 public record UpdateTransactionRequestDto(
         int id,
         String type,
-        int amount,
+        Long amount,
         String content,
         String date
 ) {
-    public UpdateTransactionRequestDto(int id, String type, int amount, String content, String date) {
+    public UpdateTransactionRequestDto(int id, String type, Long amount, String content, String date) {
         this.id = id;
         this.type = type;
         this.amount = amount;
